thumbnail-function/requirements.txt
fastapi
uvicorn[standard]
Pillow
python-multipart
python-dotenv
